improve tracer code
